Developing branch

+++ public/controller/activityManager.php
@@ -6,26 +6,24 @@ use Controller\GeneralCrud\Crud;
 $destination = "../php/activityManagement.php";
 if ($_SERVER["REQUEST_METHOD"] == "POST" && ($_SESSION['rol']==='ADM' || $_SESSION['rol']==='SAD' || $_SESSION['responsable'] === true)) {
 
-    if (isset($_GET['addNew']) && $_GET['addNew'] == 'true') {
-       
+    if (isset($_GET['addNew']) && $_GET['addNew'] == 'true') {   
         $name = Crud::antiNaughty((string)$_POST['Fname']);
         $description = Crud::antiNaughty((string)$_POST['Fdescription']);
-        $dateChckBox = $_POST['noDateSelected'];
-        
-        echo "<script>console.log('value: $dateChckBox');</script>";
-        if($dateChckBox == 1){
-            $fechaTermino = '00-00-0000';
-        }else{
+        $dateChckBox = (int)$_POST['noDateSelected'];
+        $responsable = (int)$_POST['responsableActividad'];
+        $projectID = $_SESSION['projectSelected'];
+        $idObjetivo = (int)$_POST['objetivoEnlazado'];
+
+        if($dateChckBox == 0){
             $mes = $_POST['mes_meta'];
             $dia = $_POST['dia_meta'];
             $año = $_POST['anio_meta'];
-            
+    
             $fechaTermino= "$año-$mes-$dia";
+        }else{
+            $fechaTermino = '00-00-0000';
         }
-        $responsable = $_POST['responsableActividad'];
-        $projectID = $_SESSION['projectSelected'];
-        $idObjetivo = $_POST['objetivoEnlazado'];
-        
+
         if($_SESSION['rol']==='ADM' || $_SESSION['rol']==='SAD'){
             $destination = "../php/activityManagement.php?id=$projectID";
         }else{
@@ -34,17 +32,24 @@ if ($_SERVER["REQUEST_METHOD"] == "POST" && ($_SESSION['rol']==='ADM' || $_SESSI
         $sql = "INSERT INTO `tbl_actividades` (`id_proyecto`, `id_usuario`, `nombre_actividad`, `descripción`, `estadoActual`, `participantes`, `fecha_estimada`, `id_objetivo`)  VALUES (?, ?, ?, ?, 1, null, ?, ?);";
         $params = [$projectID, $responsable, $name, $description, $fechaTermino, $idObjetivo];
         Crud::executeNonResultQuery($sql, $params, 'iisssi', $destination);
-    
     }
    
     if (isset($_POST['delete']) && $_POST['delete'] == 'true' && isset($_POST['id']) && isset($_POST['rep'])) {
-        $id = $_POST['id'];
-        $rep = $_POST['rep'];
         $project = $_SESSION['projectSelected'];
+        $destination = "../php/activityManagement.php?id=$project";
+
+        $id = filter_var($_POST['id'], FILTER_VALIDATE_INT);
+        $rep = filter_var($_POST['rep'], FILTER_VALIDATE_INT);
+        
+        if ($id === false || $rep === false || $project === false) {
+            $errorMsg = urlencode("Parámetros inválidos proporcionados.");
+            header("Location: $destination?error=$errorMsg");
+            exit();
+        }
+
         $query = "DELETE FROM tbl_actividades WHERE id_proyecto = ? AND id_actividad = ? AND id_usuario = ?";
         $params = [$project, $id, $rep];
         $types = "iii";
-        $destination = "../php/activityManagement.php?id=$project";
         Crud::executeNonResultQuery2($query, $params, $types, $destination);
     }
